requirements.txt
fastapi==0.115.5
uvicorn==0.34.0
langchain==0.3.13
langchain-community==0.3.13
chromadb==0.5.23
sentence-transformers==3.3.1
pandas==2.2.3
python-multipart==0.0.20
pydantic==2.10.3
numpy==2.2.1
aiofiles==24.1.0
supabase==2.10.0
python-dotenv==1.0.0
deep-translator==1.11.4
